Refactor: extract environment config parsing to shared layer (fixes #49)
- Created shared/config/env.ts singleton to read, sanitize, and validate ACTUAL_* environment variables.
- Removed duplicated parseJsonEnvArray, cleanEnvString, and env parsing logic from TFSAContributionsService, InvestmentProjectionsService, and DashboardSummaryService.
- Unified the hasTFSA, hasRESP, hasRRSP boolean conditions into a single source of truth.
